Move BackgroundManager to Core
