+++ app/src/main/java/io/horizontalsystems/bankwallet/core/managers/ConnectivityManager.kt
@@ -6,6 +6,7 @@ import android.net.Network
 import android.net.NetworkCapabilities
 import android.net.NetworkRequest
 import io.horizontalsystems.bankwallet.core.App
+import io.horizontalsystems.core.BackgroundManager
 import io.reactivex.subjects.PublishSubject
