+++ app/src/main/java/io/horizontalsystems/bankwallet/core/managers/EthereumKitManager.kt
@@ -6,6 +6,7 @@ import io.horizontalsystems.bankwallet.core.UnsupportedAccountException
 import io.horizontalsystems.bankwallet.entities.AccountType
 import io.horizontalsystems.bankwallet.entities.CommunicationMode
 import io.horizontalsystems.bankwallet.entities.Wallet
+import io.horizontalsystems.core.BackgroundManager
 import io.horizontalsystems.ethereumkit.core.EthereumKit
 import io.horizontalsystems.ethereumkit.core.EthereumKit.*
