+++ core/src/main/java/io/horizontalsystems/core/BackgroundManager.kt
@@ -1,4 +1,4 @@
-package io.horizontalsystems.bankwallet.core.managers
+package io.horizontalsystems.core
 
 import android.app.Activity
 import android.app.Application
